+++ web_poet/__init__.py
@@ -1,3 +1,3 @@
 from .pages import WebPage, ItemPage, ItemWebPage, Injectable
-from .page_inputs import ResponseData
+from .page_inputs import ResponseData, Meta
 from .requests import request_backend_var, Request, HttpClient